+++ src/CSVreader.h
@@ -4,10 +4,8 @@
 // simple and basic csv reader class
 
 #include <stdlib.h>
-//#include <string.h>
 #include <stdio.h>
 #include <iostream>
-//#include <list>
 #include <vector>
 
 using std::cout;
@@ -15,9 +13,6 @@ using std::wcout;
 using std::endl;
 using std::cin;
 using std::vector;
-//using std::string;
-//using std::list;
-//using std::wstring;
 
 class CCSVreader
 {
@@ -50,14 +45,14 @@ private:
 	// private functions
 	int get_header_tokens(wchar_t *p_line_cur_w, wchar_t **p_h);
 	int read_line(wchar_t **p_line_out, wchar_t **p_line_2out, int line_no);
-	int read_word(int line_no, double **pd,int index);
+	int read_word(int line_no, float **pd,int index);
 	bool isNumber(const wchar_t* str);
 	fpos_t get_pos_begining_line(int line_no);
 	int clean_mess();// close all the opened files and deallocate the memory, delete pointes
 
 public:
 	wchar_t **p_header;
-	double **p_data;
+	float **p_data;
 	errno_t err;
 
 private:
